requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-Babel==3.1.0
Flask-CORS==4.0.0
Werkzeug==2.3.7
qrcode==7.4.2
Pillow==10.0.1
python-dotenv==1.0.0
cryptography==41.0.7
opencv-python==4.11.0.86
pyzbar==0.1.92
openpyxl==3.1.2
pywebpush==1.14.0
